Renamed e2e test po file
