+++ components/settings/proposals/components/ProposalWorkflow.tsx
@@ -22,7 +22,6 @@ import { useState } from 'react';
 import { Button } from 'components/common/Button';
 import { VisibilityIcon } from 'components/common/Icons/VisibilityIcon';
 import MultiTabs from 'components/common/MultiTabs';
-import { useIsCharmverseSpace } from 'hooks/useIsCharmverseSpace';
 import { usePreventReload } from 'hooks/usePreventReload';
 import { useSnackbar } from 'hooks/useSnackbar';
 import { getDefaultEvaluation } from 'lib/proposals/workflows/defaultEvaluation';
